Extract mutation rates widget from MutationRatesDesc
Move list preview out of MutationRatesDesc into a reusable
MutationRatesWidget used by GenomeEditorWidget and MassOperationsDialog.
