+++ source/Gui/GenomeEditorWidget.cpp
@@ -16,7 +16,7 @@
 #include "GenericMessageDialog.h"
 #include "GenomeTabEditData.h"
 #include "GenomeTabLayoutData.h"
-#include "MutationRatesDialog.h"
+#include "MutationRatesWidget.h"
 #include "StyleRepository.h"
 
 namespace
@@ -94,15 +94,7 @@ void _GenomeEditorWidget::processHeaderData()
 
             AlienGui::Group(AlienGui::GroupParameters().text("Mutation rates"));
 
-            auto buttonWidth = scale(60.0f);
-            auto availableWidth = ImGui::GetContentRegionAvail().x;
-            auto listBoxWidth = availableWidth - buttonWidth - ImGui::GetStyle().ItemSpacing.x;
-            AlienGui::ListBox(AlienGui::ListBoxParameters().items(_editData->genome._mutationRates.getActiveMutationTypes()).width(listBoxWidth));
-            ImGui::SameLine();
-            if (AlienGui::Button("Edit")) {
-                MutationRatesDialog::get().open(
-                    _editData->genome._mutationRates, [this](MutationRatesDesc const& mutationRates) { _editData->genome._mutationRates = mutationRates; });
-            }
+            MutationRatesWidget::process(_editData->genome._mutationRates);
 
             AlienGui::SliderFloat(
                 AlienGui::SliderFloatParameters()
