+++ source/Gui/MutationRatesWidget.h
@@ -1,0 +1,9 @@
+#pragma once
+
+struct MutationRatesDesc;
+
+class MutationRatesWidget
+{
+public:
+    static void process(MutationRatesDesc& mutationRates, bool nested = false);
+};
